[add] input logic [edit] move player movement logic at class
