+++ ConsoleAdventure/Content/Scripts/Items/Inventory/Inventory.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using ConsoleAdventure.Content.Scripts.Player;
 
 namespace ConsoleAdventure
 {
