+++ ConsoleAdventure/Content/Scripts/Player/PlayerMovement.cs
@@ -1,0 +1,35 @@
+using ConsoleAdventure.Content.Scripts.InputLogic;
+
+namespace ConsoleAdventure.Content.Scripts.Player;
+
+public class PlayerMovement
+{
+    public bool IsMoving { get; private set; }
+    
+    private readonly Position _direction;
+    private int x;
+    private int y;
+    
+    public PlayerMovement()
+    {
+        _direction = new Position();
+    }
+
+    public Position GetDirection()
+    {
+        x = Input.GetHorizontalMovement();
+        y = Input.GetVerticalMovement();
+        
+        _direction.SetPosition(x, y);
+        
+        x = 0;
+        y = 0;
+
+        if (_direction.x != 0 || _direction.y != 0)
+        {
+            IsMoving = true;
+        }
+        
+        return _direction;
+    }
+}
