+++ ConsoleAdventure/Content/Scripts/Player/States/IPlayerState.cs
@@ -1,0 +1,6 @@
+namespace ConsoleAdventure.Content.Scripts.Player.States;
+
+public interface IPlayerState
+{
+    void HandleInput();
+}
